+++ src/user_input/mouse.rs
@@ -2,6 +2,7 @@
 
 use bevy::input::mouse::{MouseMotion, MouseWheel};
 use bevy::prelude::{MouseButton, Reflect, Resource, Vec2};
+use bevy::utils::HashSet;
 use leafwing_input_manager_macros::serde_typetag;
 use serde::{Deserialize, Serialize};
 
@@ -28,7 +29,7 @@ impl UserInput for MouseButton {
     /// as it represents a simple physical button.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Simple(Box::new(*self))
+        BasicInputs::simple(self.clone())
     }
 
     /// Creates a [`RawInputs`] from the button directly.
@@ -108,7 +109,7 @@ impl UserInput for MouseMoveDirection {
     /// [`MouseMoveDirection`] represents a simple virtual button.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Simple(Box::new(*self))
+        BasicInputs::simple(self.clone())
     }
 
     /// Creates a [`RawInputs`] from the direction directly.
@@ -193,10 +194,11 @@ impl UserInput for MouseMoveAxis {
     /// [`MouseMoveAxis`] represents a composition of two [`MouseMoveDirection`]s.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Composite(vec![
-            Box::new(MouseMoveDirection(self.axis.negative())),
-            Box::new(MouseMoveDirection(self.axis.positive())),
-        ])
+        let mut inputs: HashSet<Box<dyn UserInput>> = HashSet::new();
+        inputs.insert(Box::new(MouseMoveDirection(self.axis.negative())));
+        inputs.insert(Box::new(MouseMoveDirection(self.axis.positive())));
+
+        BasicInputs { inputs, length: 1 }
     }
 
     /// Creates a [`RawInputs`] from the [`DualAxisType`] used by the axis.
@@ -303,12 +305,13 @@ impl UserInput for MouseMove {
     /// [`MouseMove`] represents a composition of four [`MouseMoveDirection`]s.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Composite(vec![
-            Box::new(MouseMoveDirection::UP),
-            Box::new(MouseMoveDirection::DOWN),
-            Box::new(MouseMoveDirection::LEFT),
-            Box::new(MouseMoveDirection::RIGHT),
-        ])
+        let mut inputs: HashSet<Box<dyn UserInput>> = HashSet::new();
+        inputs.insert(Box::new(MouseMoveDirection::UP));
+        inputs.insert(Box::new(MouseMoveDirection::DOWN));
+        inputs.insert(Box::new(MouseMoveDirection::LEFT));
+        inputs.insert(Box::new(MouseMoveDirection::RIGHT));
+
+        BasicInputs { inputs, length: 1 }
     }
 
     /// Creates a [`RawInputs`] from two [`DualAxisType`]s used by the input.
@@ -410,7 +413,7 @@ impl UserInput for MouseScrollDirection {
     /// [`MouseScrollDirection`] represents a simple virtual button.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Simple(Box::new(*self))
+        BasicInputs::simple(self.clone())
     }
 
     /// Creates a [`RawInputs`] from the direction directly.
@@ -495,10 +498,10 @@ impl UserInput for MouseScrollAxis {
     /// [`MouseScrollAxis`] represents a composition of two [`MouseScrollDirection`]s.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Composite(vec![
-            Box::new(MouseScrollDirection(self.axis.negative())),
-            Box::new(MouseScrollDirection(self.axis.positive())),
-        ])
+        let mut inputs: HashSet<Box<dyn UserInput>> = HashSet::new();
+        inputs.insert(Box::new(MouseScrollDirection(self.axis.negative())));
+        inputs.insert(Box::new(MouseScrollDirection(self.axis.positive())));
+        BasicInputs { inputs, length: 1 }
     }
 
     /// Creates a [`RawInputs`] from the [`DualAxisType`] used by the axis.
@@ -604,12 +607,13 @@ impl UserInput for MouseScroll {
     /// [`MouseScroll`] represents a composition of four [`MouseScrollDirection`]s.
     #[inline]
     fn decompose(&self) -> BasicInputs {
-        BasicInputs::Composite(vec![
-            Box::new(MouseScrollDirection::UP),
-            Box::new(MouseScrollDirection::DOWN),
-            Box::new(MouseScrollDirection::LEFT),
-            Box::new(MouseScrollDirection::RIGHT),
-        ])
+        let mut inputs: HashSet<Box<dyn UserInput>> = HashSet::new();
+        inputs.insert(Box::new(MouseScrollDirection::UP));
+        inputs.insert(Box::new(MouseScrollDirection::DOWN));
+        inputs.insert(Box::new(MouseScrollDirection::LEFT));
+        inputs.insert(Box::new(MouseScrollDirection::RIGHT));
+
+        BasicInputs { inputs, length: 1 }
     }
 
     /// Creates a [`RawInputs`] from two [`DualAxisType`] used by the input.